Move init functions to a dedicated `env.py` file
