+++ utils/__init__.py
@@ -1,1 +1,2 @@
+from utils.env import init_logger  # For backward compatibility
 from utils.utils import *
