+++ utils/utils.py
@@ -40,27 +40,6 @@ TENANT_REGEX_HML = re.compile(r"https://.*--m3par-miia.netlify.app")
 IS_LOCAL = os.environ.get("ENVIRONMENT") == "local"
 
 
-def init_logger():
-    """
-    Should be called in the global scope of the main file.
-
-    See also: `init_env()`, which calls `init_logger()` after reading the `.env`
-    file
-    """
-    log_level = os.environ.get('MIIA_LOG_LEVEL', 'INFO').upper()
-    logging.getLogger().setLevel(log_level)
-
-
-def init_env():
-    """
-    Read `.env` files and initialize the environment and logging.
-    """
-
-    from dotenv import load_dotenv
-    load_dotenv()
-    init_logger()
-
-
 def echo_request(event):
     params = event["queryStringParameters"] or {}
